`models.py`: Add ensemble URLs for `model_seg_ms_sc_mp2rage`
NB: This release is not final, and should probably not be
merged. These initial seeded models are only to be used
for testing the new feature. We would likely want to
merge the URL changes for this model in a separate PR, once
further R&D has been done on the ivadomed side.