`models.py`: Adapt `is_valid` to work on str or list
With the introduction of model ensembles, `path_model`
can either be a single path (str) or a list of paths.

This necessitates a change in how we check for model
validity, since `is_valid` couldn't be called on
lists of paths.

The alternative to this commit would be to map the
`is_valid` function to a list of strings, using a pattern
such as:

```
valid = (all(is_valid(path) for path in path_model)
         if isinstance(path_model, list)
         else is_valid(path_model))
```

The upside here is that we preserve the simplicity of
the `is_valid` function. However, this pattern greatly
complicates the caller sites, and adds a lot of redundant
"is this a list or a string" checking. So, I went with
adapting the `is_valid` function instead.