`models.py`: Add support for model emsembles to `-list-tasks`
The two "list tasks" options, `-list-tasks` and
`-list-tasks-long`, must check for the validity of model
paths to report whether a given task is installed or not.

Given that `path_models` can now represent model ensembles,
we must add function calls to update `path_models` if it
is a model ensemble.