2nd try: Replace `_get_centerline` with `get_centerline(space="phys")`
This commit accomplishes essentially the same thing as the previous commit,
784704c4a15f24a3f9cd54c2d49cca976cf88f69, except this time, we keep the
`arr_ctl` and `arr_ctl_der` produced by `get_centerline`, and use them to
create a `Centerline` object at the caller site.

This drastically reduced the complexity of the PR, because it preserves the
existing outputs of the `get_centerline` function, meaning we don't have to
update every single call to `get_centerline`.