`install_sct`: Add `git clone` step
There are 3 possibilities for how `install_sct` has been run:
  1. git clone --> ./install_sct
  2. Download "Source code (.zip)" --> ./install_sct
  3. Download install_sct by itself --> ./install_sct

We want to only run `git clone` ourselves in case 3. So, how do we tell
if we already have the source files in the PWD?

ANS: We can re-use the existing check for the version.txt file. :)

This isn't 100% reliable, but it's hard to imagine a situation where
'spinalcordtoolbox/version.txt' isn't present but the rest of the source
isn't. Plus, this minimizes the PR diff.