`install_sct`: Move `SCT_INSTALL_TYPE` inside earlier if/else
When `install_sct` is run on its own, it can *never* be an in-place
installation, because we don't have access to the source files in the
PWD. Because of this, we must force SCT_INSTALL_TYPE=package. So, we
move the setting of SCT_INSTALL_TYPE into the earlier if/else block.