Fix flake8 lints (#4041)
* Ignore E741 in nurbs.py

The variable 'l' is just used here in the typical sequence of
generic indexing variables, 'i, j, k, l'. There's not really
a better name.

* Reformat long lines

* Remove bare `except`

By removing the entire block of code which contained it, because
it's currently buggy and not used anywhere. The `h.update` function
only accepts a bytes-like object, and throws a TypeError when
passed a string.

* Ignore bare `except`

We're going to remove sentry and raven soon anyway, so there's no
need to figure out the correct fix for this bit of code:

https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/3079

* Fix bare `except`

Clearly it's only trying to catch OSError.

* Use `printv` instead of `raise`

Python doesn't allow raising a string as an exception, so
this was just generating a TypeError instead of displaying
the intended message.

* Fix missing whitespace

* Remove bare except

This `try ... except: raise` was literally a no-op.

* Fix bare `except`

The tarfile and zipfile modules each define a bunch of custom exceptions,
so the most specific exception type we can catch here is Exception.

* Simplify exception handling

There's no need to `except Exception: raise Exception`.

* Fix typo in variable name

flake8 reported this as an unused variable, but really it
looks like a typo: the underscore should have been a period!

* Remove unused variables

As reported by flake8, these variables are assigned but the
value is never used (and doesn't seem to produce side-effects).

* Rename ambiguous variable

flake8 doesn't like the single-letter variable 'l',
presumably because it also looks like '1' or 'I'.

* 'not x in y' -> 'x not in y'

* Apply style suggestions

* Don't hide useful traceback information

---------

Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>