Remove `colormap` and use `im_types` instead
This was mostly a straightforward 1:1 swap between colormaps and im_types.

However, there was one function (sct_compute_mtsat) where I made the
decision to NOT add `im_types`. This is because both colormaps were `gray`;
in this case, we don't need to specify anything, as the viewers will default
to using the grayscale colormap if we don't specify anything.