`install_sct.bat`: Adopt `SCT_SOURCE`/`SCT_DIR`/`SCT_INSTALL_TYPE` logic
Previously, `install_sct.bat` followed this logic:

- Always treat `~/spinalcordtoolbox` as both the source dir and install dir
- Check for `~/spinalcordtoolbox/.git/` to decide whether to re-clone
- Don't distinguish between in-place installs and stable release installs

However, the Linux/macOS `install_sct` has more complicated logic:

- Have a more flexible source dir (could be PWD, could be tempdir)
- Set the install dir depending on the SCT version (`~/sct_<version>`)
- Distinguish between in-place installs and stable release installs

So, this commit takes the core logic from `install_sct` and ports it
to `install_sct.bat`, so that the steps are largely matched.

Note: This commit is missing the logic that allows the user to choose
their own installation directory. I'll add this in a separate commit.