`create-release.yml`: Fix path to `bin/`
Because the `install_sct.bat` installer now takes into account
the presence of `requirements-freeze.txt` when determining
where to install SCT, our "Create Release" tests now install
SCT to `~/sct_0.0/` for both Linux/macOS AND Windows.

This means our path to `bin/` is now incorrect; $GITHUB_WORKSPACE
is only applicable for in-place, non-release installs (i.e. tests.yml)

Additionally, on Windows, $HOMEPATH resolves to /Users/runneradmin.
But, the Windows filesystem is actually on a separate drive within
an outer unix context. So, it's not sufficient to rely purely on
$HOMEPATH when setting the $PATH variable -- we have to append a
drive letter "/d" so that it resolves to "/d/Users/runneradmin".