`sct_check_dependencies`: Set `is_sct_binary=True` in `sct_testing` call
Context: For some reason, Python isn't interpreting the `bin/`
directory on the `PATH` correctly.

Now, the `bin/` dir, as added to the `PATH`, is:

   /d/Users/runneradmin/sct_0.0/bin

Which should be evaluated to:

   D:\Users\runneradmin\sct_0.0\bin

In fact, the conversion from `/d` to `D:\` is what happens to the
old `bin/` path, `$GITHUB_WORKSPACE/bin`.

However, what the new path ACTUALLY gets evaluated is:

   /d\Users\runneradmin/sct_0.0/bin

This mangled path causes a FileNotFoundError when trying to run `sct_testing`
from within `sct_check_dependencies`.

Ideally we would try to fix this at the source, by setting `PATH` in a way that
causes Python to do the `/d`->`D:\` conversion correctly. But, as a quick and dirty
fix, setting `is_sct_binary=True` causes `run_proc` to use an `absolutepath` for
`sct_testing`, which completely removes any dependence on how Python interprets the
`PATH` variable.

Since this is really just a GHA-specific issue for `windows-2019` runners,
and not an issue that affects user installs, I think it's okay to use this
fix as a long-term solution.

(And, anyway, `sct_testing` *is* an SCT binary, so we probably should have
been using this option from the get-go.)