`testing/`: Fix "`E402`: module level import not at top of file"
More info: https://www.flake8rules.com/rules/E402.html

Author's note: These errors are largely caused by the statement:

> sys.path.append(os.path.join(__sct_dir__, 'scripts'))

I think this is a relic from when SCT used to include its scripts
outside the `spinalcordtoolbox` package directory, and when it
called scripts via a subprocess.

But, now that `scripts/` is a module itself, and we call our
scripts in-process, this PATH modification is no longer necessary.

So, we can probably just get rid of those statements entirely? But,
for now, I've simply addressed the linting issue.