Replace `-setorient-data` with `-flip` and `-transpose` options (#4071)
`-setorient-data` is essentially a way to perform flips and
transposes on an image's data array. However, its usage requires the
user to specify an orientation string *relative to the orientation
string already in the header*, which has historically been confusing
to our users.

To make things clearer, we can instead provide flip and transpose
functionality directly to our users, so that they explicitly know
what operations they are applying to their data.