Workaround for 'undefined name' linting errors
flake8 emits F821 linting errors for some objects which are provided
by the FSLeyes context which runs this code. By doing a no-op
re-definition of these objects of the form "x = x", we can inform
flake8 (and pycharm, etc.?) that we expect these variables to be
defined, without writing '# noqa: F821' comments everywhere.