`sct_label_utils`: Add weak support for duplicate labels
This isn't perfect, because the checkboxes act weirdly, but
at least the basic functionality is there (can select and
undo labels -- just not update existing labels).

To make this work 100% perfectly, we would need to dig into
`base.py` to find out what logic is used to check and uncheck
the checkboxes. Frankly, I don't quite understand it, so I'm
leaving this as a TODO, because this won't be necessary until
SCT v6.1.