Remove unused `verbose` parameters
These functions don't use the `verbose` parameter at all, so there's
no need to pass it in. Since there are no callers outside the file,
it's a local change.

The function `register_step_dl_multimodal_cascaded_reg` also seems to
have an unused `verbose` parameter, but it creates a `Param()` object
with a hardcoded `'verbose': 0` value, so maybe it **should** be used,
so I didn't remove it.