Align default value for `verbose` with actual usage
Some of these functions have a default of `verbose=1`, and others
have `verbose=0`, which is confusing. What's more, these functions
are always called with a value for `verbose`, so the default is
never used. The actual source for this parameter is all the way
up the call chain, in `sct_register_multimodal.py`, where the
actual default is 1.

I would rather not set a default value at all if it's never used,
but in this case that would case an error:
```
SyntaxError: non-default argument follows default argument
```
This could probably be fixed in a bigger code cleanup.