Use `sct_progress_bar` for `algo=columnwise` and `register2d` slicewise loops (#4094)
* `algorithms.py`: Use `sct_progress_bar` instead of `logger.info`

This matches `columnwise` and `register2d` with the
usage of `sct_progress_bar` in `algo=centermassrot`.

* `algorithms.py`: Add newlines before `sct_progress_bar`

This preserves the existing logging spacing. :)

(There might be a better way to do this than `print()`? But,
 adding `\n` to tqdm's "desc" parameter doesn't work.)

* Respect verbosity settings

In the case of register2d, verbosity is lower because the command is
called inside an sct_progress_bar loop.

* Remove unused `verbose` parameters

These functions don't use the `verbose` parameter at all, so there's
no need to pass it in. Since there are no callers outside the file,
it's a local change.

The function `register_step_dl_multimodal_cascaded_reg` also seems to
have an unused `verbose` parameter, but it creates a `Param()` object
with a hardcoded `'verbose': 0` value, so maybe it **should** be used,
so I didn't remove it.

* Align default value for `verbose` with actual usage

Some of these functions have a default of `verbose=1`, and others
have `verbose=0`, which is confusing. What's more, these functions
are always called with a value for `verbose`, so the default is
never used. The actual source for this parameter is all the way
up the call chain, in `sct_register_multimodal.py`, where the
actual default is 1.

I would rather not set a default value at all if it's never used,
but in this case that would case an error:
```
SyntaxError: non-default argument follows default argument
```
This could probably be fixed in a bigger code cleanup.

* `algorithms.py`: Lessen `verbose` in register2d `run_proc` call

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guaypaquet@gmail.com>