`Slice.mosaic()`: Convert into an `abstractmethod`
Before this commit, `mosaic()` was an instance method that belonged to the more
general `Slice` class. This implies that `mosaic` should work as-is for _all_
`Slice` classes (i.e. `Axial` AND `Sagittal`).

However, `mosaic` was designed specifically for the `Axial` class, and doesn't really
work all that well for the `Sagittal` class. So, it doesn't really make sense for it
to be an instance method for the parent `Slice` class; it should instead be an
abstract method, with implementations being specific to Axial/Sagittal.

(NB1: There are a few TODO comments scattered throughout `Slice` that say:
         TODO : Move this to the Axial class
 So, I think this could be part of a larger effort to refactor the Slice classes.)

(NB2: It may be possible to write a mosaic method that works for both Axial and Sagittal
 slices. But, for now, to ensure that I don't affect the existing Axial mosaics, I'm
 going to work on the Sagittal implementation separately from the Axial implementation.)