`cropping.py`: Refactor `ImageCropper.crop` to accept images
Previously, `ImageCropper` could only be used to crop a single
image (`self.img_in`) per instance. But, for the Sagittal QC,
we want to crop many slices using an identical bounding box each
time.

So, this commit adds a quick workaround to allow `crop()` to
accept images, overriding the `self.img_in` parameter.

NB: In the long run, it would probably be more straightforward
if there wasn't a `self.img_in` attribute in the first place.
That way, `ImageCropper` instances would be less limited in their
ability to crop multiple images. But, doing this would require a
much larger refactor, and I wanted to preserve backwards
compatability the best I could within the scope of this PR.