`Sagittal.mosaic`: Use `ImageCropper` + SC seg to crop images
((NB: This commit cuts out *many* axial-specific details.))

Re-using the `ImageCropper` class was the easiest way to
determine the bounding box from an input segmentation.

However, by using this approach, I've abandoned the `Slice`
class's convention of implementing a `self.crop` method. So,
while this is a working proof of concept, it should probably
be refactored further such that all slice types follow a similar
convention.

If anything, I would suggest that the Axial class also adopt
`ImageCropper`. But, this can be discussed further in review.