Change Axial/Sagittal to axial/sagittal for `-plane` argument
The lowercase versions are easier to type.

This change was prompted specifically by the new `-plane`
argument in `sct_qc`. However, the `generate_qc` function
was already using `plane` internally (with capitalized
'Axial' and 'Sagittal' strings).

So, we could either:

- Change *every* instance internally to lowercase, or
- Accept lowercase arguments in the CLI script, but
  capitalize them before passing them on to internal
  functions.

I went with the latter option to keep the changes in this
PR minimally-scoped. But, it would be cleaner in-code to
bite the bullet and change everything to lowercase (perhaps
in a separate PR for organizational purposes?).