`create-release.yml`: Use actual version instead of dummy 0.0 version
Because we want to test the "git clone" standalone installation path,
we will be git cloning the source code to a temporary directory,
and thus it will be trickier to inject a dummy value into version.txt.

The whole reason for using a dummy value was that it was supposed to
replicate the Release Step where version.txt is updated. But, I don't
think this really matters in terms of testing? Using e.g. 6.0.dev0
isn't all that different than using 0.0. I can't imagine a scenario
where 0.0 would reveal a bug and 6.0.dev0 wouldn't.