`create-release.yml`: Induce standalone install by copying to $HOME
This is just a test for this PR only, to see if a standlone install
even works using these steps.

In reality, the copying/cd'ing should only happen on master, because
otherwise, standalone installations would break for PRs from forks.

So, once we confirm that these steps work, we can add some sort of
conditional, such as:

        if [ "${GITHUB_REF#refs/heads/}" == "master" ]; then