`.ci.sh`: Try standalone installation for UNIX platforms
In `tests.yml`, UNIX platforms use `.ci.sh` as a way to centralize
the repeated steps across various OSs. So, I have to make my changes
here instead of `tests.yml`.

This time, I'm using `/Downloads` to try to fully replicate the user
experience, and avoid issues with working directly in the home directory.

If this works on the PR, I will add an `if` to make sure these new steps
only run on master.