Add sagittal mosaic QC and use it for `sct_deepseg_lesion` QC report (#4102)
* Optimize imports

* Implement QC for axial sct_deepseg_lesion.

- add new 'sct_deepseg_lesion' choice for '-p' flag
- add '-plane' flag.

* Add example command for sct_deepseg_lesion

* `qc.py`: Add Sagittal QC option for `sct_deepseg_lesion`

This is non-functional until I implement the mosaic method.

* `Slice.mosaic()`: Convert into an `abstractmethod`

Before this commit, `mosaic()` was an instance method that belonged to the more
general `Slice` class. This implies that `mosaic` should work as-is for _all_
`Slice` classes (i.e. `Axial` AND `Sagittal`).

However, `mosaic` was designed specifically for the `Axial` class, and doesn't really
work all that well for the `Sagittal` class. So, it doesn't really make sense for it
to be an instance method for the parent `Slice` class; it should instead be an
abstract method, with implementations being specific to Axial/Sagittal.

(NB1: There are a few TODO comments scattered throughout `Slice` that say:
         TODO : Move this to the Axial class
 So, I think this could be part of a larger effort to refactor the Slice classes.)

(NB2: It may be possible to write a mosaic method that works for both Axial and Sagittal
 slices. But, for now, to ensure that I don't affect the existing Axial mosaics, I'm
 going to work on the Sagittal implementation separately from the Axial implementation.)

* `Sagittal.mosaic()`: Use annotated copy of `Axial.mosaic()`

This will need to be heavily edited. But, by adding annotations,
I'm able to better understand each aspect of the Axial implementation
of the function, so that I know which areas to change.

(In theory, these annotations could be ported back to `Axial.mosaic`,
 but again, I want to leave `Axial.mosaic` untouched for now.)

* `Slice.add_slice()`: Remove `size` parameter

This was done for two reasons:

1. `size` is unnecessary. We can infer the size from `patch.shape`. In fact,
   we already do exactly this elsewhere in the `add_slice()` function!
2. The single variable `size` assumes a square patch. This is only true for Axial
   slices, and is not true for Sagittal slices. So, this change will allow for
   rectangular patches, and thus will help the Sagittal implementation of `mosaic`.

* `cropping.py`: Refactor `ImageCropper.crop` to accept images

Previously, `ImageCropper` could only be used to crop a single
image (`self.img_in`) per instance. But, for the Sagittal QC,
we want to crop many slices using an identical bounding box each
time.

So, this commit adds a quick workaround to allow `crop()` to
accept images, overriding the `self.img_in` parameter.

NB: In the long run, it would probably be more straightforward
if there wasn't a `self.img_in` attribute in the first place.
That way, `ImageCropper` instances would be less limited in their
ability to crop multiple images. But, doing this would require a
much larger refactor, and I wanted to preserve backwards
compatability the best I could within the scope of this PR.

* `Sagittal.mosaic`: Use `ImageCropper` + SC seg to crop images

((NB: This commit cuts out *many* axial-specific details.))

Re-using the `ImageCropper` class was the easiest way to
determine the bounding box from an input segmentation.

However, by using this approach, I've abandoned the `Slice`
class's convention of implementing a `self.crop` method. So,
while this is a working proof of concept, it should probably
be refactored further such that all slice types follow a similar
convention.

If anything, I would suggest that the Axial class also adopt
`ImageCropper`. But, this can be discussed further in review.

* `Sagittal.mosaic`: Fixup outdated docstring info

* `cropping.py`: Change `if not img` to `if img is None`

Avoids corner case where img_in is not None, but == None.

* `qc.py`: Deduplicate `plane` cases for lesion QC

* `slice.py`: Replace `ceil(x // y) + 1` with `ceil(x / y)`

You can either do floor division and +1, or float division and ceil, but there's no need to do both at the same time. I went with the more direct option (ceil).

* `slice.py`: Replace `int(x / y)` with `(x // y)`

These both accomplish the same thing (truncating the decimals, converting to int) but using floor division is a little clearer about intent in this specific context, since `int` implies a type conversion.

* Change Axial/Sagittal to axial/sagittal for `-plane` argument

The lowercase versions are easier to type.

This change was prompted specifically by the new `-plane`
argument in `sct_qc`. However, the `generate_qc` function
was already using `plane` internally (with capitalized
'Axial' and 'Sagittal' strings).

So, we could either:

- Change *every* instance internally to lowercase, or
- Accept lowercase arguments in the CLI script, but
  capitalize them before passing them on to internal
  functions.

I went with the latter option to keep the changes in this
PR minimally-scoped. But, it would be cleaner in-code to
bite the bullet and change everything to lowercase (perhaps
in a separate PR for organizational purposes?).

* `slice.py`: Turn `mosaics_through_time` into an `Axial`-specific method

This allows us to improve the function signatures of the various
`mosaic` calls, as it allows us to keep the `return_center` parameter
as an Axial-specific parameter.

* `slice.py`: Remove `dilation` from `Sagittal.mosaic`

This is a useful feature to have in theory, but has issues in its
interaction with `ImageCropper` that would need to be rectified
in order for this to be properly implemented:

https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/4102#discussion_r1194357002

* `slice.py`: Unify signatures of `mosaic` methods

In our calls to `mosaic` in the codebase, the optional
parameters `size` and `nb_columns` were never actually
used.

So, by inlining these parameters, we can simplify the
method signatures, and make each subclass compatible
with the parent class.

* `sct_qc.py`: Remove `metavar=PLANE`

`metavar=PLANE` was blocking the `choices` from being shown.

Before:

  -plane PLANE          Plane of the output QC. Only relevant for -p sct_deepseg_lesion.

After:

  -plane {axial,sagittal}
                        Plane of the output QC. Only relevant for -p sct_deepseg_lesion.

---------

Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>