`slice.py`: Compute 0/1 % using slices to avoid test failure
In one of our tests, we warp the PAM50 template, then use the warped
PAM50 image files in QC. One of the warped image files is completely
blank, meaning it contains only one unique value (0).

But, to compute the percentage, we currently use count[0] and
count[1], which naturally fails because there's only one value.

Using slicing gets around this, since counts[0:2] will succeed
even if there is only one value in the array.

(This won't be considered a segmentation (as expected), because segs
still require both 0 *and* 1 to be the most common values.)