`resampling.py`: Introduce option to preserve s/qform codes
This the second of two causes of the q/sform codes being reset to defaults:

https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/3005#issuecomment-1404341719

By making this an option that defaults to False, we can opt in case by case at caller sites for `resample_nib`. This ensures backwards compatibility, while also making sure that devs *really* intend to preserve the codes in specific instances.