`install_sct`: Use explicit paths to python/pip executables (#4143)
* `install_sct.bat`: Use explicit paths to python/pip executables

The `conda` executable already used an explicit path, and so it didn't need to be changed.

* `install_sct`: Use explicit paths to python/pip/conda executables

* `install_sct`: Remove `python` activation check

The only real purpose of activating a conda environment is to do some PATH manipulation to prioritize conda's executables.

But, since we add SCT's `/bin` folder to the path, there isn't ever a need to activate the conda environment from the user's perspective.

So, even if the conda environment is activated incorrectly, it won't matter, since we just
access the `python` executable directly. This prevents us from having to debug activation
issues in cases where SCT would have installed correctly anyway.

* `install_sct`: Remove now-unused realpath polyfill

* `install_sct`: Revert explicit path to `conda` executable

This isn't actually necessary like I thought it was. Adding the explicit path doesn't change whether the correct package is installed into the SCT environment, since that instead hinges on whether the environment was activated correctly.

* Specify explicit path to python

To be consistent with the other calls to python and pip.

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guaypaquet@gmail.com>