Fix checkboxes by using idx, not val for `self.check_boxes`
While this commit does seem to work in testing, it touches several existing attributes:

- base.py :: BaseController (in undo(), self.points: [x, y, z, label] -> [x, y, z, label, idx])
- sagittal.py :: SagittalController.points ([x, y, z, label] -> [x, y, z, label, idx])
- widgets.py :: VertebraeWidget.check_boxes (dict -> list)
- widgets.py :: VertebraeWidget.check_boxes (QCheckBox.label -> QCheckBox.val)
- widgets.py :: VertebraeWidget.label (setter method: set by index, not value)
- widgets.py :: VertebraeWidget.label (getter method, return QCheckBox object, not value)
- widgets.py :: VertebraeWidget.labels (setter method, set by value using QCheckBox.val)

This worries me, because it may be possible that some existing usages of these attributes were missed, and thus other QC functionality was inadvertently affected by these changes.

(For example, does changing BaseController.undo() affect the CenterlineController (in addition to the SagittalController? Do CenterlineController's `self.points` need to be updated to include an index, too?)