`sct_register_to_template`: Copy cache file to output dir
Previously, the `cachefile` path made it so that `cache_save()` saved the file to `curdir` (i.e. the working directory).

After this change, `curdir` is only used to check whether an *input* cache file exists. But, for saving, we now A) save to the temporary directory and B) copy it to the output directory.