Refactor `get_closest` to avoid confusing nested abs/rel calls
Previously, `get_closest_to_absolute` was calling `get_closest_to_relative` and returning early, which was quite confusing! It's also misleading, since the return value from `absolute` won't actually be `absolute` most of the the time.

To fix this, I've refactored this method so that we first call a centralized `get_closest_index` method, which *then* calls either `relative` or `absolute` depending on the position of the vertebral_level.

This vastly simplifies the `absolute` function, which helps us to untangle its logic in isolation (which will be greatly useful for diagnosing https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4085).

----

NOTE: There is a slight change to the logic concerning how `Centerline.regions_labels` is accessed. Before, it was accessed from both the `backup` and `self` centerlines. Now, it's accessed from just `self`. The reason I made this change is that `regions_labels` is a class attribute of the `Centerline` class, and thus is not data-dependent, so it will not differ between the two centerlines.