`types.py`: Revert fix from https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/4192
The check only really mattered when the space of `backup_index` differed from the space of `backup_centerline`. But, this was actually a bug!! The spaces should be the same. I just didn't realize it at the time of authoring #4192.

Now, with the fix from the previous commit, the two spaces will always be identical, i.e. backup_index will always exist inside backup_centerline. So, the check has been made redundant.