`sct_deepseg`: Inline `options` dictionary
This section of code will turn into a conditional, where we use different functions to segment the image depending on whether the model comes from ivadomed, monai, or nnunet.

Because of this, I wanted to make it clearer which lines of code were "ivadomed-only" and which lines of code were relevant to `sct_deepseg` as a whole. By inlining the options dict, we isolate the ivadomed-specific code, leaving us more free to add non-ivadomed code.