`types.py`: Rewrite convoluted logic for clarity
This code is critical because it identifies when the uppermost disc label is *not* C1, which is relevant when labeling the points above the uppermost disc label. So, I want to make sure that this code is readable and well-documented.

Before, there was this odd "upper = 31" logic, essentially going from bottom to top to identify the uppermost label. But, I don't see why we can't just take the `min` of the available labels instead.

NB: It's possible that we don't even need to do this? There are some steps earlier on in the function where `self.first_label` is identified. And like... `self.first_label` is basically the uppermost label!! What I don't understand quite yet is the case where `self.first_label` would never be assigned, so I wasn't quite comfortable merging the two sections of code.