`types.py`: Ensure that `self.first_label` is always assigned + reuse it
This is a follow-up commit that directly addresses the extended message in my previous commit. (Merging two sections of code, such that "uppermost label" = `self.first_label`)

`index_first_label` and `index_last_label` can only be `None` if _none_ of the labels in `discs_levels` are present in `self.list_labels`.

If such a case were to occur, then there would be further errors later on down the function, because we would have no valid disc labels!! Surely we should halt the function right away and alert the user.

As an added bonus, since we can be sure that `self.first_label` is properly set, then we can access it directly when setting the reference label.