`inference.py`: Rework multiclass splitting logic to use `dataset.json`
Rather than hardcoding the label names and their associated values, we can read this information directly from the `dataset.json` file.

This works for both the SCI model _and_ the mouse GM/WM model! And, it removes the need to infer logic from the dataset name, as we were doing for the SCI model.