`image.py`: Mitigate precision issues due to `float32` zooms
Context:

- https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4378#issuecomment-1953004537
- https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4378#issuecomment-1953033814

Because the `dim` property is initially stored as `float32`, when we perform multiplication using
the dim values, some spurious 1e-9 digits are added to the computed CSA values.

To fix this, we can compute the the "absolute degree of discretization" of the values using
np.spacing. This tells us what we can safely round to without introducing error.

This fix is relatively minor in terms of numerical accuracy, but it's very important from an
"understanding" perspective, because it ensures that the uncorrected CSA values have relatively
few significant figures. Meaning, the neat "evenness" of the summed CSA will demonstrate the limits
on precision for low-mm resolution images.