`qc.py`: Make color list generation more robust to edge cases
The existing code works as follows:

- We start with a list of labels. This QC was designed for vertebral labels, so typically this might be something like [2, 3, 4, 5, 6] (corresponding to C2, C3, C4, etc.)
- We then create a list of N colors, where N is the range from the minimum label (2) to the max label (6).
- The colors are sourced from a base list of 7 high contrast colors, repeated.

However, confusingly the base list of 7 colors is repeated by hardcoding a `* 15` multiplier. Meaning, if we try to fetch colors from this list using label values greater than `7 * 15 = 105`, the resulting behavior is unclear. (Will an error be thrown? Will we get an incomplete list?)

To handle this case, we can use a more robust method for "repeating the list of colors until we have N colors" that properly scales to the number of labels.

This change alone is enough to fix the issue described in #4379, however it doesn't produce the prettiest results. I will try to improve this further in a separate commit.

(Side note: I see some other issues with the existing "draw text" code, but I'll tackle those separately!)