`qc.py`: Prettify label coloring by parsing distinct label groups
The implementation of this is up for debate (see the discussion in issue #4379). But, the gist of this change is:

- The previous method (repeating high contrast colors) works OK for vertebral labels, but it doesn't work well for a complex set of labels with multiple semantically-related groups (cord/CSF, vertebral levels, intervertebral discs), as using a single color palette for all of these groups makes them difficult to distinguish visually.
- So, we try to automatically detect the different groups, then assign a unique colormap to each of them, while still preserving the "usual" behavior when there is only 1 continuous group.

Note that this could perhaps be simplified if we knew for certain that we would always be dealing with the same label values (e.g. 200 = spinal cord, 201 = CSF). But, I wanted to try to implement a more "general" solution, since we won't necessarily be able to predict what labels our uses will try to visualize.