Add new QC option `-exclude-label-text` for vertebral labels
Note 1: There's an inversion here between the CLI argument ("exclude text") vs. the API method property ("draw text"). I figure that "exclude text" makes sense for a CLI argument, because text is drawn by default, so users must opt out of that behavior. But, on the API level, "if not exclude_text:" is a really strange pattern to grok. So, I figure a good place to invert this logic is at the interface between CLI and API (`draw_text=(not arguments.exclude_label_text)`).

Note 2: This changeset isn't ideal to me; what I really wanted to do was pass an argument directly to the `label_vertebrae` method. However, with how the current QC reports are designed, this isn't possible -- parameters must be passed to the `QcImage` class, stored in the object, then accessed by the method.

This problem is fixed in Mathieu's QC revamp prototype (#4224), as each QC-generation method is simplified down to a basic function whose only job is to create two PNG images. With such a structure, a hypothetical `label_vertebrae` function could accept a parameter for whether to draw text on the image.

(Hypothetically, we could also employ subparsers one day to better contain each process-specific argument such as -exclude-label-text and -fps, too.)