Switch to using mean magnitude for moco QC
The purpose of this QC file is to identify the amount of motion in each volume
(that is, at each time point), so that motion-related fluctuation in the signal
can be corrected for.

Having the average of the signed X and Y values is not useful for this, because
a lot of left-right motion (or front-back motion) could mostly cancel out, and
look the same as not having much motion at all. This is potentially confusing,
and prone to misuse.

In contrast, the mean magnitude is only small when there is not much total
motion, and it is scale-invariant with the number of Z slices.