Pin to `dipy<1.6` until conflicts with `dipy>=1.8` can be resolved (#4398)
## Description

`dipy==1.8` and above cause conflicts with SCT's venv due to ivadomed.
More details here: https://github.com/ivadomed/ivadomed/pull/1308

So, for now, we need to pin `dipy<1.6` to avoid breakage every time a
new version of `dipy` is released. But, once
https://github.com/ivadomed/ivadomed/pull/1308 is merged and a new
version of ivadomed is released, we can switch this pin to
`dipy!=1.6.*,!=1.7.*` and allow 1.8/1.9/etc.

## Linked issues

Fixes #4397.