Move "Create Release" CI workflow to be cron-only
This change reduces the number of concurrent macOS runners from 3 to 2. Given the overall limit of
5 macOS runners, this is an important change, since it effectively allows us to push twice to PRs
concurrently without exceeding our quota. (2x2=4 vs. 2x3=6)

I targeted this workflow because, in most cases, this will be identical to running tests normally.
(The same test suite is run as tests.yml, and usually with the same dependencies, too!)

There are some exceptions though. The biggest one is when we change our requirements.txt file in a
way that isn't cross-OS compatible. So, long-term, I think it would be a good idea to try to
conditionally re-enable this test on PRs under certain circumstances (e.g. if `requirements.txt`
has been modified). But, in the short term, as we work to unblock the current glut of PRs, I
figured it would be OK to go for the simplest approach.

(See https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/3669 for more context as to
why the create-release.yml was added in the first place.)