Restore `macos-14`/`windows-2019`/`ubuntu-22.04` to Release .yml
On reflection, because the create-release.yml job is now cron-only, it makes sense to try to test
a wider range of OSs, since we don't really incur additional costs for doing so.

Plus, since Create Release is specifically meant to test installation incompatibilities, especially
cross-platform ones due to `pip freeze`, I think it makes an extra amount of sense to enable
additional OSs for this workflow.