Replace DOI medRxiv link with direct medRxiv link
For some reason, the DOI link is breaking, even though the underlying medRxiv link loads fine.
The broken DOI link occurs even on my personal browser, leading me to believe that this isn't an
issue with the newly-added workflow.

I don't think there's a meaningful difference between one or the other, so I'd rather just
replace the link and get this PR merged?